+++ sportgems/exc.py
@@ -1,0 +1,19 @@
+
+class DistanceTooSmallException(Exception):
+    pass
+
+
+class InconsistentLengthException(Exception):
+    pass
+
+
+class TooFewDataPointsException(Exception):
+    pass
+
+
+class NoSectionFoundException(Exception):
+    pass
+
+
+class InvalidDesiredDistanceException(Exception):
+    pass