+++ docs/changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [0.6.0](https://github.com/fgebhart/sportgems/releases/tag/0.6.0) - 2021-MM-DD
+
+### Changed
+* All sportgems python exceptions have been moved to the dedicated
+  `sportgems.exceptions` submodule.
+
+
 ## [0.5.0](https://github.com/fgebhart/sportgems/releases/tag/0.5.0) - 2021-04-03
 
 ### Changed